Put back splashes into the document flow
Previsouly, the splashes (front page, download page, and a few releases pages)
were positioned using 'position: absolute', which meant that other elements on
the page could overlap with those splashes. For this reason, the first element
after the front spash page had a 20em top margin. This was fragile and caused
this elements to be hidden by the splash on small devices.

This commit puts back the splashes into the document flow. Splashes backgrounds
have a right margin of the size of the menu, and are now Block Formatting
Contexts, which allows overlays to be placed in absolute position but still
staying inside the splash.

This is a first step towards a more responsive www.gimp.org website. On
desktops, this results in the same website: this is on purpose!

Signed-off-by: Michael Schumacher <schumaml@gmx.de>