Modified front page sections a bit to make it cleaner.
I've moved the "STILL WORKING" column out of the section with the
intro blurb (which I've also revised a little bit).  This is to
start looking more critically at the sections and verbiage as
part of a final push on the front page specifics.

If anyone wants to take a stab at any of the sections text, please
please, please do!