Explain *how* to contribute to the GIMP projects.
The "Development Status", "Bug Reports" and "Web Development" headings seemed to
me like they were really subsections of "How You Can Help," so I set those to be
h3 headings.

A couple times, in the "Hacking" and "How To Submit Your Changes" sections, I
tried to borrow the tone of Wikipedia's "Be Bold" guidelines.
<https://en.wikipedia.org/wiki/Wikipedia:Be_bold>

I added a list of links to the most important GIMP repositories. I was going to
mention gimp-web-devel in this list, but it looks like no one has pushed any
commits there since 2013. Perhaps that repo should be archived?

Signed-off-by: Pat David <patdavid@gmail.com>