Added different 'Donation' buttons to header navigation.
We used to have a simple Flattr banner/button there, but it left
off bitcoin and paypal from visibility.

I have replaced the Flattr badge with a PayPal, Flattr, and
bitcoin icon (with appropriate links) that will now be visible
on all pages of the site in the header navigation.