Modify index.html template to show news/author list
index.html is the base template for indexing articles (news)
authors.html extends index.html with some blocks.  I updated
the index.html blocks to correctly replace the page title with
the author name if it's being extended by authors.html (and to
keep saying "Recent News" if it's not).