Fix search bar to use duckduckgo if no js
JS is a requirement for the tipue static search on the site
to work.  IF the use has disabled JS, default to sending
them to searching the site using duckduckgo.

There's an intermediate problem becaues the non-js version
of duckduckgo doesn't accept a site= parameter in it's
query string :\.  They expect you to munge the site= param
with the query itself.  Which we can't do without resorting
to JS, which we wouldn't have to do IF JS was enabled in
the first place (and so on and so forth...).

So, if no JS - duckduckgo will be used.  The user wil be
sent to ddg, which will then send them to the non-js version
of _their_ site.  At least this way to site= param stays
intact.