Remove status: hidden from metadata for download page
I had previously set the Status: hidden attribute on
this page to hide it on some default themes.  This is
not required on the final WGO theme, and has the side effect
of hiding the page from getting parsed for the search.

I also moved all of the styling tags/entries to the end
of the article to remove it from the search preview.