Add statistics on contribution.
I think that's cool to show the vibrant activity around GIMP but also to
thank our contributors. I tried to not forget anyone, and in the same
time not to appear like we are overflowing with contributors (which we
are definitely not). Feel free to update or even delete if you think
this section is too much.