Fix image scaling issue
Reported by prokoudine

I had set images to have a max-height: 100vh;
(100% of the viewport height).  This way any images that are too
tall can be seen in their entirety.

The problem is I had also set the image width: 100%;
this scaled the images funny (they would squish if they were too
tall, but maintain full width).

This commit keeps the image max-height: 100vh; and removes the
width: 100% so the image can resize as needed to viewport height.