Bug 773364 - Wilber SVG icon is black.
Mozilla Firefox ignores inline style rules in SVG because of Content
Security Policy. Exporting a SVG in Inkscape as Optimized SVG replaces
CSS rules with XML attributes when possible. As a second step, I deleted
all style="" attributes with a search and replace. This seems to be
working fine with simple SVG images and the optimized file is also a lot
smaller.
I kept around the original in Inkscape SVG format though, to avoid data
loss. I also took the opportunity to add the appropriate authorship
metadata for Aryeom in the SVG files.