content: get rid of the embedded data image in the 404 animated SVG.
I was embedded a piece of the banner ("Niagara Rainbow" by Ville Pätsi)
from the landing page as the header which Wilber is trying to pull.
Chrome is ok, but Firefox handles badly this embedded data on testing.
The possible reason is CSP, since it works well locally. Anyway let's
just replace it with a unicolor rectangle.