content: improving the 404 SVG animation for broken Firefox output.
Firefox seems to have problem on animating the visibility XML attribute
when a CSP policy is in action (likely case since it works locally, but
not on testing).
Is it a bug? Or is this XML attribute considered as style?
Fix various usages by replacing them with translation out of canvas or
opacity animation (not sure if opacity animation will work with CSP.
Let's see).
There is likely more to fix, but I want to see how it goes so far.