Redo random homepage header image
I redid the randomized header image on the homepage.
Thanks to drc for the idea to just have the randomization happen
on build, and to schumaml for actually making that happen every
hour.

There's now a header-images.json file in the root of the project.
Add new images to be included in that file - all of the entries
are mandatory please.

I've pushed the actual randomization and inclusion in the home.css
to a plugin for Pelican, random_header.  Look there for any
of that code.

Otherwise the actual setting is now created at the end of the
pelicanconf* files.  You'll see them there.

Thank you drc and schumaml!