Set pelicanconf.local.py to use a cache
Previously, pelicanconf.local.py was deleting the build directory
before building the site.  This was mainly for testing purposes
while designing the site, and is not really needed now.

I've set it to _not_ delete the build directory automatically.
I've also set it to use the cache to speed up subsequent builds -
this should be helpful when writing new pages and not wanting to
wait as long for a rebuild to check things.