Remove go-parts.com mirrors from mirrors2.json
The UK and RU mirrors for go-parts are redirecting to an
auto-parts domain?  The main go-parts is unresolved.
The US one works, but the previous three have me worried.

I've removed them for now.  If they are actually good
mirrors still, then we can just revert this commit to get
them back in.